Support using as library
